# back-ogs/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OGS",
        "contact": {
            "email": "mickpicart@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "OGS API"
        }
    ],
    "paths": {
...
        "/loaderrors/{id}": {
            "put": {
                "tags": [
                    "Errors"
                ],
                "summary": "Load one website errors to DataBase by Id",
                "description": "Load one website errors to DataBase by Id",
                "operationId": "loadOneWebsiteErrorsInDb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The supervision_datas id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/loadextdataserrors/{id}": {
            "put": {
                "tags": [
                    "External Datas and Errors by Id"
                ],
                "summary": "Load external datas and errors to DataBase for one website",
                "description": "Load external datas and errors to DataBase for one website",
                "operationId": "loadOneWebsiteExtDatasAndErrorsToDb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The website id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/loadextdataserrorsall": {
            "put": {
                "tags": [
                    "All External Datas and Errors"
                ],
                "summary": "Load external datas and errors to DataBase for all websites",
                "description": "Load external datas and errors to DataBase for all websites",
                "operationId": "loadAllWebsiteExtDatasAndErrorsToDb",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/supervisiondbdatas": {
            "get": {
                "tags": [
                    "Supervision DB Datas"
                ],
                "summary": "Get all supervision datas from DataBase",
                "description": "Returns supervision datas from DataBase",
                "operationId": "supervisionDBDatas",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
        "/website": {
            "get": {
                "tags": [
                    "Websites"
                ],
                "summary": "Get websites list",
                "description": "Returns websites list from DataBase",
                "operationId": "websitesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Websites"
                ],
                "summary": "Register a new website",
                "operationId": "websiteRegister",
                "parameters": [
                    {
                        "name": "url",
                        "in": "query",
                        "description": "The website url",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Website successfully registered",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "409": {
                        "description": "Resource already exists"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}